fix(knapsack): docsDir base path /docs => /docs-dir
This is for the older docs approach where config.docsDir was a path to a folder of markdown files. These were viewable at /docs/filename. The newer approach of custom sections is much preferred and setting a custom section for Docs at the /docs base path is desired, which conflicted with what docsDir used. This change allows /docs to be used by Custom Sections.